Move and rename datasource/helper to /src folder
